Commited Database Connection and Combining Customer table to datatable (getCustomers).

+++ CustomerProject/CustomerProject/ViewModels/Customer.cs
@@ -7,8 +7,7 @@ namespace CustomerProject.ViewModels
 {
     public class Customer
     {
-
-        public Customer(String Name, int Age, String Address, int PhoneNumber, String Gender)
+        public Customer(String Name, int Age, String Address, long PhoneNumber, String Gender)
         {
             this.Name = Name;
             this.Age = Age;
@@ -20,7 +19,7 @@ namespace CustomerProject.ViewModels
         public String Name { get; set; }
         public int Age { get; set; }
         public String Address { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public String Gender { get; set; }
 
     }
